Add download endpoint, simplify error handling, move out of modules
